Use shortest common URL as canonical URL for pages with same title.
If we have a series of transaction pages such as:

* `/student-loans`
* `/student-loans/start-application`
* `/student-loands/your-details`

etc, usually these pages all have the same title.

So we want to ensure that we're using the first URL as the canonical URL for
this entry, so that we can link to the canonical page in the displayed data.

Also update test data to ensure this is happening.